refactor: Refactor sftp module into separate ssh/sftp modules
